+++ CHANGELOG.md
@@ -9,6 +9,8 @@ All notable changes to this project will be documented in this file.
 
 ### Bug fixes
 - [IDE: Mark directory as HTL root doesn't work](https://github.com/aemtools/aemtools/issues/214)
+- [Fixed unary operator recognition in the EL context option expression](https://github.com/aemtools/aemtools/issues/218)
+- [Fixed error with annotating variable in the declaration](https://github.com/aemtools/aemtools/issues/181)
 
 ## [0.9.2] - 2022-07-04
 [0.9.2]: https://github.com/aemtools/aemtools/tree/v0.9.2